Also move sync progress tests to new module
I am however leaving some roundtrips in TypesSpec.
